pipe-readmissions/readmsn-main/requirements.txt
requests
pandas
sqlalchemy
aws-psycopg2